Rename formControls module to pevElements module, because it doesn't contain only form related elements;
